+++ ShopGo/CartAnimationFiles/CartAnimationStoryboard.storyboard
@@ -11,7 +11,7 @@
         <!--Cart Animation View Controller-->
         <scene sceneID="s0d-6b-0kx">
             <objects>
-                <viewController storyboardIdentifier="cartAnimationViewController" id="Y6W-OH-hqX" customClass="cartAnimationViewController" customModule="ShopGo" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="CartAnimationViewController" id="Y6W-OH-hqX" customClass="CartAnimationViewController" customModule="ShopGo" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="5EZ-qb-Rvc">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
